Fix type mismatches and implement simple distance-based location matching
## Fix Type Mismatches and Implement Distance-Based Location Matching

**Problems fixed:**
1. Type mismatch in technician model: changed `f32` scores to `f64` to match PostgreSQL FLOAT8 storage
2. Removed PostGIS dependency: spatial.rs was using unavailable `ST_MakePoint` and `ST_DWithin` functions
3. Implemented Haversine formula for distance calculation in memory

**Changes:**
- Updated `src/models/technician.rs`: changed `certification_level`, `part_authenticity_score`, and `speed_score` from `f32` to `f64`
- Rewrote `src/services/spatial.rs`:
  - Implements Haversine distance formula for accurate lat/lon calculations
  - Fetches all verified technicians and filters by distance in memory
  - No external dependencies required

**Result:**
- Technician creation now works (no type mismatch)
- Location matching now works without PostGIS
- All 6 API endpoints should now be fully functional

+++ backend/src/models/technician.rs
@@ -8,12 +8,12 @@ pub struct Technician {
     pub id: Uuid,
     pub user_id: Uuid,
     pub shop_name: String,
-    pub location_name: String, // e.g., "Mega City, Kisumu", "Milimani"
+    pub location_name: String,
     pub latitude: f64,
     pub longitude: f64,
-    pub certification_level: f32, // e.g., 1.0 for Apple Certified, 0.7 for Independent Expert
-    pub part_authenticity_score: f32, // e.g., 0.0 to 1.0 rating based on supply chain ledger
-    pub speed_score: f32, // Turnaround efficiency score
+    pub certification_level: f64,
+    pub part_authenticity_score: f64,
+    pub speed_score: f64,
     pub is_verified: bool,
     pub created_at: DateTime<Utc>,
 }
@@ -25,9 +25,9 @@ pub struct CreateTechnicianRequest {
     pub location_name: String,
     pub latitude: f64,
     pub longitude: f64,
-    pub certification_level: f32,
-    pub part_authenticity_score: f32,
-    pub speed_score: f32,
+    pub certification_level: f64,
+    pub part_authenticity_score: f64,
+    pub speed_score: f64,
 }
 
 #[derive(Deserialize)]
@@ -36,3 +36,4 @@ pub struct MatchQueryRequest {
     pub longitude: f64,
     pub max_distance_km: Option<f64>,
 }
+
